Time: 152 ms (37.87%), Space: 71 MB (20.42%) - LeetHub

+++ 3066-minimum-operations-to-exceed-threshold-value-ii/3066-minimum-operations-to-exceed-threshold-value-ii.java
@@ -1,17 +1,17 @@
 class Solution {
     public int minOperations(int[] nums, int k) {
-        PriorityQueue<Integer>minHeap=new PriorityQueue<>();
+        PriorityQueue<Long>minHeap=new PriorityQueue<>();
         int n=nums.length;
 
         for(int i=0;i<n;i++){
-            minHeap.add(nums[i]);
+            minHeap.add((long)nums[i]);
         }
         int count=0;
 
         while(minHeap.size()>=2 && minHeap.peek()<k){
-            int x=minHeap.poll();
-            int y=minHeap.poll();
-            int result=(Math.min(x,y)*2 + Math.max(x,y));
+            long x=minHeap.poll();
+            long y=minHeap.poll();
+            long result=(Math.min(x,y)*2 + Math.max(x,y));
             minHeap.add(result);
             count++;
         }
